Party Follow fixed!
I've finally found a smooth model for this.  Basically I'm creating a route on the fly, the movement controller uses it just like any other route.

Still needs work for Zone changes.

I still need to go through assist and things like that to refine it and add more options.

git-svn-id: http://pocketgnome.googlecode.com/svn/trunk@389 4a9c585c-cb2d-11dd-a3c7-fda619b15ae3