This has a long list of bug fixes, some were introduced with new features, some were legacy bugs.
Added 2 new whisper commands follow and stop.

Redid weighting.

Added a lot of notifications for error handling and new error handling.

There's a lot more in this commit, but it's been a weeks worth of work and I've not kept notes on it all... just trying to get this out.

Be sure to check my 1.5 features list here;
http://pg.savorydeviate.com/forum/viewtopic.php?f=64&t=4356

This release is pretty stable, there are still issues we'll find I'm sure, but it's at a point where it's worth the commit.

git-svn-id: http://pocketgnome.googlecode.com/svn/trunk@496 4a9c585c-cb2d-11dd-a3c7-fda619b15ae3