Movement controller was redone, many conflicts that resulted from botController vs movementController are gone.
Casting, we now check to see if spells are on CD or usable, also check item CD's.  In theory you should now only need to use the spell cd condition if you really have to.

UI, I added some buttons similar to the button next the the Profile pulldown.  I also made it so that the buttons pull up their respective objects.

git-svn-id: http://pocketgnome.googlecode.com/svn/trunk@511 4a9c585c-cb2d-11dd-a3c7-fda619b15ae3